requirements.txt
absl-py==0.10.0
astunparse==1.6.3
autopep8==1.5.4
beautifulsoup4==4.9.1
blis==0.4.1
boto==2.49.0
boto3==1.14.49
botocore==1.17.49
cachetools==4.1.1
catalogue==1.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cssselect==1.1.0
cycler==0.10.0
cymem==2.0.3
docutils==0.15.2
gast==0.3.3
gensim==3.8.3
google-auth==1.21.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.31.0
h5py==2.10.0
idna==2.10
importlib-metadata==1.7.0
jmespath==0.10.0
joblib==0.16.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
llvmlite==0.34.0
lxml==4.5.2
Markdown==3.2.2
matplotlib==3.3.1
murmurhash==1.0.2
nltk==3.5
numba==0.51.0
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
pandas==1.1.1
Pillow==7.2.0
pkg-resources==0.0.0
plac==1.1.3
preshed==3.0.2
protobuf==3.13.0
pt-core-news-lg==2.3.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
readability==0.3.1
readability-lxml==0.8.1
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
s3transfer==0.3.3
scikit-learn==0.23.2
scipy==1.4.1
seaborn==0.10.1
six==1.15.0
smart-open==2.1.1
soupsieve==2.0.1
spacy==2.3.2
srsly==1.0.2
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
thinc==7.4.1
threadpoolctl==2.1.0
toml==0.10.1
tqdm==4.48.2
umap-learn==0.4.6
Unidecode==1.1.1
urllib3==1.25.10
wasabi==0.7.1
Werkzeug==1.0.1
wordcloud==1.8.0
wrapt==1.12.1
zipp==3.1.0
